modified:   src/controllers/store/customer.controller.js 	modified:   src/routes/store/orderRoute.js 	modified:   src/services/order.service.js 	new file:   src/utils/mapOrderStatus.js
